+++ src/main/java/com/intellij/plugins/haxe/lang/psi/HaxeGenericSpecialization.java
@@ -135,9 +135,9 @@ public class HaxeGenericSpecialization implements Cloneable {
           HaxeResolveResult resolved =
             HaxeResolveResult.create((HaxeClass)context, fromGenericResolver(context, holder.getClassType().getGenericResolver()));
           specialization.put(element, name, resolved);
-        } else if (holder.getClassType() != null) {
+        } else if (holder.getClassType() != null && !holder.isUnknown()) {
           HaxeClass clazz = holder.getClassType().getHaxeClass();
-          HaxeResolveResult resolved = HaxeResolveResult.create(clazz, fromGenericResolver(context, holder.getClassType().getGenericResolver()));
+          HaxeResolveResult resolved = HaxeResolveResult.create(clazz, fromGenericResolver(null, holder.getClassType().getGenericResolver()));
           specialization.put(element, name, resolved);
         } else if (holder.getFunctionType() != null) {
           SpecificFunctionReference type = holder.getFunctionType();