+++ src/main/java/com/intellij/plugins/haxe/model/type/HaxeExpressionEvaluator.java
@@ -564,6 +564,13 @@ public class HaxeExpressionEvaluator {
           }
         } else {
 
+
+          // unwrap Null
+          //TODO make util for unwrap/get underlying type of Null<T>? (or fix resolver ?)
+          if (typeHolder.getType().isNullType()) {
+            typeHolder = typeHolder.getClassType().getSpecifics()[0];
+          }
+
           // TODO: Yo! Eric!!  This needs to get fixed.  The resolver is coming back as Dynamic, when it should be String
 
           // Grab the types out of the original resolver (so we don't modify it), and overwrite them
@@ -630,6 +637,16 @@ public class HaxeExpressionEvaluator {
               }
             }
           }
+          else if (subelement instanceof HaxeFieldDeclaration fieldDeclaration) {
+            if(fieldDeclaration.getVarInit() != null) {
+              HaxeVarInit init = fieldDeclaration.getVarInit();
+              HaxeExpression initExpression = init.getExpression();
+              HaxeGenericResolver initResolver = HaxeGenericResolverUtil.generateResolverFromScopeParents(initExpression);
+              typeHolder = HaxeTypeResolver.getFieldOrMethodReturnType((AbstractHaxeNamedComponent)subelement, initResolver);
+            }else if (fieldDeclaration.getTypeTag() != null){
+              typeHolder = HaxeTypeResolver.getTypeFromTypeTag(fieldDeclaration.getTypeTag(), fieldDeclaration);
+            }
+          }
           else if (subelement instanceof HaxeMethodDeclaration methodDeclaration) {
             SpecificFunctionReference type = methodDeclaration.getModel().getFunctionType(resolver);
             typeHolder = type.createHolder();
@@ -682,6 +699,12 @@ public class HaxeExpressionEvaluator {
           }
 
 
+          else if (subelement instanceof HaxeSwitchCaseCaptureVar caseCaptureVar) {
+            HaxeSwitchStatement switchStatement = PsiTreeUtil.getParentOfType(caseCaptureVar, HaxeSwitchStatement.class);
+            if (switchStatement.getExpression() != null) {
+              typeHolder = handle(switchStatement.getExpression(), context, resolver);
+            }
+          }
           else if (subelement instanceof HaxeSwitchCaseExpr caseExpr) {
             HaxeSwitchStatement switchStatement = PsiTreeUtil.getParentOfType(caseExpr, HaxeSwitchStatement.class);
             if (switchStatement.getExpression() != null) {
@@ -1016,6 +1039,19 @@ public class HaxeExpressionEvaluator {
             }
           }
           return left.getArrayElementType().getType().withConstantValue(constant).createHolder();
+        }
+        //if not native array, look up ArrayAccessGetter method and use result
+        if(left instanceof  SpecificHaxeClassReference classReference) {
+          HaxeNamedComponent getter = classReference.getHaxeClass().findArrayAccessGetter(resolver);
+          if (getter instanceof  HaxeMethodDeclaration methodDeclaration) {
+            HaxeMethodModel methodModel = methodDeclaration.getModel();
+            HaxeGenericResolver localResolver = classReference.getGenericResolver();
+            HaxeGenericResolver methodResolver = methodModel.getGenericResolver(localResolver);
+            localResolver.addAll(methodResolver);// apply constraints from methodSignature (if any)
+            ResultHolder returnType = methodModel.getReturnType(localResolver);
+            if (returnType.getType().isNullType())localResolver.resolve(returnType);
+            if (returnType != null) return returnType;
+          }
         }
       }
       return SpecificHaxeClassReference.getUnknown(element).createHolder();